+++ CheckedExceptions/AnalyzerSettings.cs
@@ -22,6 +22,12 @@ public partial class AnalyzerSettings
     [JsonIgnore]
     internal bool IsLinqImplicitlyDeclaredExceptionsEnabled => !DisableLinqImplicitlyDeclaredExceptions;
 
+    [JsonPropertyName("disableLinqEnumerableBoundaryWarnings")]
+    public bool DisableLinqEnumerableBoundaryWarnings { get; set; } = false;
+
+    [JsonIgnore]
+    internal bool IsLinqEnumerableBoundaryWarningsEnabled => !DisableLinqEnumerableBoundaryWarnings;
+
     [JsonPropertyName("disableControlFlowAnalysis")]
     public bool DisableControlFlowAnalysis { get; set; } = false;